requirements.txt
langgraph==1.0.3
langchain-community==0.4.1
pypdf==6.2.0
langchain-chroma==1.0.0
langchain-openai==1.0.2
python-dotenv==1.2.1
langchain==1.0.5